added a tray icon

+++ FinishedProduct/Startup.py
@@ -15,14 +15,25 @@ from PyQt5.QtWidgets import QApplication
 
 machines_file_path = 'machines.csv'
 
-def create_vault_file():
+def create_vault_file_and_REMT_folder():
     file_path = 'C:\\ProgramData\\.Vault1851320.txt'
-    try:
-        with open(file_path, 'w'): 
-            pass
-        os.system('attrib +H "{}"'.format(file_path))
-    except Exception as e:
-        print(f"Error occurred while setting file attributes: {e}")
+    folder_path = 'C:\\ProgramData\\REMT'
+
+    # Check if REMT folder exists, create if it doesn't
+    if not os.path.exists(folder_path):
+        try:
+            os.makedirs(folder_path)
+            print(f"Folder '{folder_path}' created successfully.")
+        except Exception as e:
+            print(f"Error occurred while creating folder: {e}")
+    if not os.path.exists(file_path):
+        try:
+            with open(file_path, 'w'):
+                pass
+            os.system(f'attrib +H "{file_path}"')
+            print(f"File '{file_path}' created and hidden successfully.")
+        except Exception as e:
+            print(f"Error occurred while setting file attributes: {e}")
 
 def create_machines_file():
     """Check if machines.csv file exists; if not, create an empty file."""
@@ -50,7 +61,7 @@ class FirstLoginWindow(QtWidgets.QWidget, Ui_FirstLoginForm):
         self.setupUi(self)
         self.MasterPWDContinue.clicked.connect(self.save_password)
         create_machines_file()  # Ensure machines.csv is created if it doesn't exist
-        create_vault_file()
+        create_vault_file_and_REMT_folder()
 
     def save_password(self):
         password = self.PasswordInput.text()
